[1/n - 67] 헤더 검색창 태그 필터 기능 (#150)
* ✨ searchResult, searchDetail atom 생성

* ✨ searchAtom postId key 추가

* ✨ Header 검색 결과 창 추가

* 🎨 SpanWrapper export로 변경

* 🎨 mealItem UI 수정

* ✨ SearchItem 컴포넌트 생성

* ✨ SearchDetail 컴포넌트 생성

* ✨ meal/atom 생성

* ✨ comment/atom 생성

* ♻️ home 페이지 atom 적용하여 refactoring

* ♻️ sideBar 컴포넌트 atom 적용하여 refactoring

* ♻️ MealList 컴포넌트 atom적용하여 refactoring

* ♻️ MealItem 컴포넌트 atom 적용하여 refactoring

* ♻️ MealDetail atom 적용 및 refactoring

* ♻️ DetailInfo 컴포넌트 MealDetail에서 분리

* ♻️ DetailMenu 컴포넌트 MealDetail 컴포넌트에서 분리

* ♻️ MealDetailFooter 컴포넌트 MealDetail 컴포넌트에서 분리

* ✨  선택 tag에 따라 카테고리 filter 및 mealListEntry 업데이트

* 🐛 filter로 제외된 방이 filter 해제 시 복구되지 않는 bug fix
