+++ src/component/home/meal_item.jsx
@@ -2,15 +2,25 @@ import React from "react";
 import styled from "styled-components";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { solid } from "@fortawesome/fontawesome-svg-core/import.macro";
+import { useRecoilState, useRecoilValue, useSetRecoilState } from "recoil";
+import {
+  detailOpenAtom,
+  postIdAtom,
+  userStatusAtom,
+} from "../../recoil/meal/atom";
+import { SpanWrapper } from "./meal_detail";
+import { FlexRowDiv } from "../header/Header";
 import { useRecoilState, useSetRecoilState } from "recoil";
 import { postIdAtom, userStatusAtom } from "../../recoil/meal/atom";
 import { SpanWrapper } from "./meal_detail";
 import { FlexRowDiv } from "../header/Header";
 import RequestStateTag from "../history/RequestStateTag";
 
-const MealItem = ({ meal, onMealClick, ToggleDetailBar, isDetailOpen }) => {
+const MealItem = ({ meal, onMealClick, ToggleDetailBar }) => {
   const setUserState = useSetRecoilState(userStatusAtom);
+  const isDetailOpen = useRecoilValue(detailOpenAtom);
   const [postId, setPostId] = useRecoilState(postIdAtom);
+
   return (
     <LiWrapper
       style={{ cursor: "pointer" }}
