+++ src/recoil/meal/atom.js
@@ -9,3 +9,31 @@ export const userStatusAtom = atom({
   key: "userStatusAtom",
   default: "",
 });
+
+export const mealListEntryAtom = atom({
+  key: "mealListEntryAtom",
+  default: [],
+});
+export const originalMealListEntryAtom = atom({
+  key: "originalMealListEntryAtom",
+  default: [],
+});
+export const selectedMealAtom = atom({
+  key: "selectedMealAtom",
+  default: {},
+});
+
+export const detailOpenAtom = atom({
+  key: "detailOpenAtom",
+  default: false,
+});
+
+export const openAtom = atom({
+  key: "openAtom",
+  default: false,
+});
+
+export const openIdAtom = atom({
+  key: "openIdAtom",
+  default: "",
+});
